added delete project

+++ crowdsorting/app_resources/sorting_algorithms/ACJProxy.py
@@ -111,5 +111,7 @@ class ACJProxy:
 
     def delete_self(self):
         print(f"deleting {self.acj} pickles")
-        os.remove(f"crowdsorting/app_resources/sorter_instances/{self.project_name}.pkl")
-        shutil.rmtree(self.logPath)
+        if os.path.exists(f"crowdsorting/app_resources/sorter_instances/{self.project_name}.pkl"):
+            os.remove(f"crowdsorting/app_resources/sorter_instances/{self.project_name}.pkl")
+        if os.path.exists(self.logPath):
+            shutil.rmtree(self.logPath)
